AER3-345 Generating source based on a yaml file
The .yaml file is placed in the /src/main/resources/static/ folder so spring mvc will ensure it gets served as static content.
Using a specific servlet path to include versioning in it (not too sure if this is the correct spot for it, but for now it'll do)
Moved from jersey to spring mvc for the REST part: While I do like the possibility for basic jax-rs annotations the generated code just wasn't as useful. It did not have the delegate pattern and I could not get to a clear method of implementing the service. Generated code seemed to be meant to be altered manually which just isn't as useful.
As a bonus the error messages that the application sends back are much nicer now: with jersey it was just a tomcat error page with the status code.  With the spring approach we get a json error object which seems more useful (especially since we can influence if messages are added to the response by using the server.error.include-message property).