AER3-345 Updated yaml to have dataset in path instead of query parameter
Also renamed 'version' to 'dataset' throughout.
Not yet implemented, but added a method to obtain information about different datasets as well (for now just an array of strings containing possible values).
Slightly annoying part of using dataset in the path is that every call will now end up in DatasetsResource. It might get big if we implement all emission factor categories. Guess we can delegate the calls to other beans if it gets too big...