AER3-345 datasets in database and JOOQ generated classes (#5)
* Using schema-per-dataset approach in database
Template schema is introduced, which can be cloned (using {import_common_into_schema}) for other datasets. This will work as long as the structure stays pretty much the same, if structure ever changes either old data has to be altered or a new template has to be introduced.
Public can contain values that should stay the same throughout the different datasets (like substances).
i18n tables specific for datasets have been moved so they are now prefixed with `i18n_` (bit like a schema within a schema). Other option would've been for a separate i18n dataset schema I guess, but we want to keep these together.
Moved the enum types for emission factors to a separate folder/module, since we don't want to copy these along with every new dataset. These shouldn't change as much either (if they do change a lot it should probably be implemented as a table).

* Generated JOOQ classes from database
Using the generated classes in the existing SectorRepository. So far it seems overkill, but once we start implementing stuff like farms this should save us from typo's and such.
Added a readme explaining how to re-generate this classes when database is changed.

* Setting dataset on a per-call basis
The JOOQ settings object can be used to replace the template scheme by the actual dataset scheme that should be used.
These Settings don't seem as threadsafe however (need to supply it when creating DSLContext and it's mutable), so we had to do some threadlocal approach.
That is, as long as we don't use other threads (some reactive programming anyone?) this approach should work OK.

Renamed 'version' to 'dataset', but this is a WIP (more to come when datasets get a table to query from and the openapi yaml file is changed again).

* generated JOOQ classes are generated in other source folder
Should be a bit clearer that these files are generated this way and should not be altered manually. Another benefit is that the packages won't clutter the source folder that the developer does want to change.

* Using template module
Sadly we can't use {import_common} within this module since the CommentCollector can't seem to handle this properly.
Importing a dataset will have to be done in a separate file per dataset this way, otherwise the build script will detect some recurring/circular issue (since it can't tell that these files are used for different schemas). Having a file per dataset seems fine I guess.