AER3-345 Added table to contain dataset information
This should be used to determine the different datasets and their respective schema's. Could perhaps do this based on available schema's and filtering out some, but that makes it harder to add some metadata like which one is the LATEST.
Idea is that dataset_code is used by users, and that the schema_name does not necessarily have to be the same. Could be useful when there is need for a dataset identification that isn't necessarily a good schema name (or an existing one like 'public').