AER3-345 Implemented fodder measures
Slightly changed API: The reduction fractions are per substance, and if one is available all should be available. For that reason it makes sense to have it in one object instead of separate EmissionReductionFraction objects.
Fodder measures are actually based on animal. It might make sense to add a specific call for an animal category, where the animal category would contain an array of lodging categories and an array of fodder measures