AER3-345 openapi extending (#6)
* Updated yaml to have dataset in path instead of query parameter

Also renamed 'version' to 'dataset' throughout.
Slightly annoying part of using dataset in the path is that every call will now end up in DatasetsResource. It might get big if we implement all emission factor categories. Guess we can delegate the calls to other beans if it gets too big...

* Adjusted FarmRepository to not use TEMPLATE directly

* Implemented some more api calls
Retrieving nested objects seems a bit harder with JOOQ unless you use a RecordMapper (or I'm missing something). For now opted for a second roundtrip to retrieve emission factors.

* Extended api yaml for everything farm lodging related
Using inheritance for all Category-like objects.
Sector is now Category-like (and ID is deprecated in favor of code. Kept around since it's an integer which is also used in IMAER).

* Added table to contain dataset information
This should be used to determine the different datasets and their respective schema's. Could perhaps do this based on available schema's and filtering out some, but that makes it harder to add some metadata like which one is the CURRENT.
Idea is that code is used by users, and that the schema_name does not necessarily have to be the same. Could be useful when there is need for a dataset identification that isn't necessarily a good schema name (or an existing one like 'public').

* Added all API methods to resource that we still need to implement

* Implemented call to obtain a list of possible datasets
Implemented the dataset behaviour for validating a dataset and getting the CURRENT dataset based on database as well.
Did add 'CURRENT' as a valid dataset. Seemed OK, but guess there can be arguments for not adding it to the list since it is not actually a dataset code.

* Using query parameter as a search filter
Meaning that the animal code supplied won't have to be an exact match, but can be just 'A' as well.