AER3-345 implementing farm lodging (#7)
* Implemented calls for additional lodging systems
Implemented the database retrieval part for obtaining additional lodging system information.

* Implemented reductive systems
No longer using the _VIEW methods. These are also joined to system definitions, and since you can have multiple system definitions that ends up with multiple rows with same emission factor.
Might be able to do something like an entity collector with JOOQ, but haven't found out how to do that yet.

* Added scrubber and system definitions to additional and reductive systems
Overlooked these properties when looking at AERIUS-II (missed that there is a a shared base class for these system categories).

*  Added extra properties to specific farm lodging call

* Implemented fodder measures
Slightly changed API: The reduction fractions are per substance, and if one is available all should be available. For that reason it makes sense to have it in one object instead of separate EmissionReductionFraction objects.
Fodder measures are actually based on animal. It might make sense to add a specific call for an animal category, where the animal category would contain an array of lodging categories and an array of fodder measures

* Using DbUtil in DatasetRepository
And using stream to determine the correct languagecode. Don't think it's an improvement in performance (well maybe since there was no break before, but it's only 2 values anyway), but it's a bit more readable.

* Fixed some sonar issues
Using sonar plugin for eclipse with a binding to the aerius sonar instance. Some issues popped up.

* Avoiding some code duplication