Avoid executing openapi plugin during docker dependency step
The build would break in docker because it would expect the api.yaml to be available (and that is copied later)
Were 2 solutions as far as I could see: adding the api.yaml or preventing the need for it. Since in this step we only want to download dependencies so we don't have to do that each time this seems the cleaner approach.