AER-2219 Units inland shipping EF in DB
Updated documentation for units for inland shipping emission factors. Change was already done on AERIUS-II.
Values are not yet read from database, but should be taken into account when designing the actual service for this part. (in AERIUS-II these are converted to kg/(m * ships) and kg(hour*ships) resepectively when reading from DB, but here it might make more sense to keep the database units).