AER3-428 Added a reference to the traditional lodging system if available
This reference is sometimes needed to get to the correct emission factor: If a lodging system without scrubber has a emissionfactor smaller than 0.3 * the emission factor of the linked traditional lodging system, and other reductive systems are used that DO include a scrubber, the emission factor is constrained to be 0.3 * the emission factor of the linked traditional lodging system. Take note that the webservice currently does NOT explain this.