Ensuring errors are handled properly
As it was, a circular error would appear, because we've mapped everything under /api/v1/, which is directly used within a Controller.
As that controller did not have something mapped for '/error', an error would occur, and that in turn would try to lookup something mapped for '/error', which...
Also means we can skip the whitelabel configuration.